split test files per component
